cody: Don't scroll to bottom when scrolled up in transcript (https://github.com/sourcegraph/sourcegraph/pull/50008)
I found it pretty annoying that the cody UI keeps scrolling down while
it's still producing new content. That makes it very hard to start
reading longer messages before they're done, or reading previous ones /
copying code from them. So I asked Cody for how to detect the scrolling
position and made it so that it doesn't scroll anymore when the user
scrolled up a bit.