cody: include context for last interaction with non-empty context messages (https://github.com/sourcegraph/sourcegraph/pull/50381)
Currently, we include context for the last message only. That works fine
in most cases, except when a user asks a follow-up question, and we
detect we don't have to include context. In that case, we would like to
include the context from the previous question (the one we want to
follow up on).

For example:

1. I ask, "How do access tokens work in Sourcegraph?", Cody fetches a
bunch of context messages and produces an answer.
2. I ask, "How can I delete one?", and we detect no context is needed.
Rather than including no context messages, we keep the context messages
from 1.

Also, see the test added in
`client/cody-shared/src/chat/transcript/transcript.test.ts`.

Additionally, I increased the number of context messages we request from
codebase context. The idea is to include as much context as possible
without pushing the prompt token limits.

## Test plan

* Manual testing.

## App preview:

-
[Web](https://sg-web-rn-cody-include-context-for-last.onrender.com/search)

Check out the [client app preview
documentation](https://docs.sourcegraph.com/dev/how-to/client_pr_previews)
to learn more.