focus cody textarea when webview is opened and/or focused (https://github.com/sourcegraph/sourcegraph/pull/50681)
This makes it faster to use Cody with just the keyboard. The user can
immediately start typing to Cody after invoking `Cody: Focus on Chat
View` with the keyboard.


## Test plan

Run extension in debug mode. Hit Opt+/ to open webview. The textarea
should be focused. Click back to a file. Hit Opt+/ to focus the webview.
The textarea should be focused again.