show cody context files as an action in the transcript (https://github.com/sourcegraph/sourcegraph/pull/50691)
Instead of showing these in a longer message after Cody's response, the
files that Cody read are now shown in a small chip before the response:
`7 files read`. Expanding it shows the first step (searching) and each
file. This UI can be extended to more kinds of plugins/actions in the
future.


![image](https://user-images.githubusercontent.com/1976/232328027-1426d280-d528-4563-9ca6-661fae89ff61.png)

![image](https://user-images.githubusercontent.com/1976/232328045-52aadf35-7a57-4c57-b74e-09f9a69fc055.png)



## Test plan

Open Cody. Run a query that requires it to read files. Confirm they show
up here.