cody: Add error handling for non-2xx status codes (https://github.com/sourcegraph/sourcegraph/pull/50917)
This adds an additional branch that doesn't attempt to read until a done
event is received and instead just reads the body to completion and then
calls the onError callback. After this change, API errors at least
appear as the notice that was supposed to fire. We should properly
integrate this into the UI and not just as a toast message, but that can
be done separately.