Cody: Add jaccard-based context to inline completions (https://github.com/sourcegraph/sourcegraph/pull/51161)
This brings the inline completion suggestions quality up to parity with
multiline suggestions by adding the same jaccard-based context based on
open editor tabs and the editor history into the prompt.

To avoid the two implementation diverging again in the future, I've also
put some of the logic into an `abstract class`. Unfortunately the diff
is hard to read because I also moved out the completion providers into
their own file now.

Behavior for multiline completions should be unchanged.

While working on this, I had a few really good inline completions from
Cody already 🎉 This is starting to get somewhere!

## Test plan

I added a bunch of `console.log`s to the prompt creation code to make
sure the snippets that are added to the knowledge library are
reasonable. This is getting really annoying to debug, so a better way to
log which prompts are created is starting to get more interesting IMO.



https://user-images.githubusercontent.com/458591/234633770-7a703720-84a9-43f2-bea3-07c2715c07c0.mov



<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->