Cody: Stop trying to escape HTML to make it visible (https://github.com/sourcegraph/sourcegraph/pull/51797)
Closes https://github.com/sourcegraph/sourcegraph/pull/51776

Oof, this has been an uphill battle so far. 

## The (not so happy) history

To give some context, it all started with this issue: https://github.com/sourcegraph/sourcegraph/pull/50680

Cody was passing-through unescaped HTML with side effects. The most
annoying and visible ones in these issue were `<input>` and `<style>`.

After some investigation, we learned that this is because of an overly
permissive DOMPurify config: Basically all non-XSS-related HTML tags
were allowed in the output, including `<input>` and `<style>`.

We started exploring different route for fixes and I decided to go down
the route of escaping non-code parts of the markdown so that `<input>`
would be converted to `&lt;input&gt;` and would keep being visible to
the user.

This took a couple of back and forth to get right and we still haven't
figured out a way to make it match the rendered output. One persisting
issue, for example, is that the Markdown lexer returns a different
structure for a given text then the final parser result -- And that is
with us already using the same engine! This caused me to add workarounds
like waiting for the whole markdown block to start the escaping since
the issue was particularly visible when code blocks were started but not
ended yet.

However, even after various iterations on this, we kept getting reports
of issues.

There are also some existing patterns that make this even harder:

- We are injecting HTML into the markdown before we go through any of
the markdown conversation/escaping. This needs to happen on the vscode
thread (so before we're even in the browser thread) since that's how we
render hallucination <span>s or the error messages. -> This prevents us
from doing these conversations in the react components that would fix
the next issue:
- We apply the escaping on any input, including the user input, and
directly to the response of the LLM. This means that if a user pasts
`You & Me` in the chat, they're were actually sending `You &amp; Me` to
the LLM. This would then cause more unexpected behavior because the LLM
might respond with the escaped characters.

Most of these workarounds were added so that if Cody responds with
`<input/>` it would render `<input/>` on the screen, and not just remove
it. Given the complexity of this issue though and the current situation
of having html escaping artifacts visible randomly, I’m not sure if this
is a maxima that we should strive forward to in the short term anymore.

Unforutnalty there is also no way to configure DOMPurify to escape tags
that are not matching the config. It will always remove them.

## This PR

In this PR, I’m reverting everything from the escaping journey and move
from a denylist approach to an allowlist for DOMPurify.

This makes sure that only HTML attributes that we need for the markdown
rendering will end up in the final HTML.

### How does this affect UX

- If you ask Cody to write HTML **and don't use markdown to format it**,
tags will be removed now
- If you ask about a HTML tag in the chat and do not wrap it inside a
code block, they will not be visible in the UI (but the LLM will receive
the right string)

## Test plan

Here are some examples:

<img width="578" alt="Screenshot 2023-05-11 at 14 59 32"
src="https://github.com/sourcegraph/sourcegraph/assets/458591/8b6cb916-857f-4a14-a962-7501d0ffe30e">
<img width="560" alt="Screenshot 2023-05-11 at 14 57 41"
src="https://github.com/sourcegraph/sourcegraph/assets/458591/92e490ed-8af6-4a31-89d8-9af618687898">
<img width="596" alt="Screenshot 2023-05-11 at 14 56 53"
src="https://github.com/sourcegraph/sourcegraph/assets/458591/890b9f2f-cba2-47a2-92cc-05a710f266b2">


<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->