cody: guardrails debug command (https://github.com/sourcegraph/sourcegraph/pull/52146)
This provides some basic infrastructure for guardrails that will be
built on top of. Initially the only command exposed is a debug command
which will send off the selected code for attribution. The
implementation of attribution is naive. The intention is to get some a
basic shape of how things fit together to be further improved.

Test Plan: run cody with debug and experimental.guardrails turned on.
Then select some code and run the guardrails debug command. Observe
hits. Then create some unique looking text and run attribution against
that. Get no hits.