Cody completions: Initial experimentation with multiline inline completions (https://github.com/sourcegraph/sourcegraph/pull/52205)
Part of https://github.com/sourcegraph/sourcegraph/pull/51344 

This PR adds initial exploration for inline completion of multiple
lines. The approach applied here does some very rudimentary language
parsing (we might need to look into treesitter for this to enable
support for more languages). The rules are as follows:

To trigger multiline completion:
- We must be at an empty line
- The line must be the beginning of a new block (i.e the last character
of the prefix is an `{`).

To find the right place to cut off the completion:
- We capture the initial indentation of the first line of the completion
- We include everything up to the point where the indentation goes
_below_ our current start
  - However we do include the last line if it is a sole `}`
- We skip over lines starting with `} else ` so we can chain all
conditions in an if block
  - We skip over empty lines as they do not have an indentation.

@beyang I’m curious what your thoughts on this are?  

## Test plan



https://github.com/sourcegraph/sourcegraph/assets/458591/a224db92-ef77-4e5c-b4db-4e833d59d2aa



<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->