Cody: always fetch context on first chat interaction (https://github.com/sourcegraph/sourcegraph/pull/52348)
We've had a few bug reports where Cody doesn't fetch context for a chat
question and produces a bad response. This is because we have logic to
detect when context is needed vs. when it's a follow-up question, and
it's hard to get this right with 100% accuracy.

This PR updates the chat recipe to always fetch context on the first
interaction. **This prevents a really bad user experience where on their
first chat question, Cody refuses to respond and says it needs more
information about the codebase.** Even if context isn't needed for the
question, it's not harmful since LLMs tend to be good at ignoring
irrelevant content.

## Test plan

Added new unit test. Manually tested that this fixes the reported issue.