cody: guardrails aggregate attribution from dotcom and private (https://github.com/sourcegraph/sourcegraph/pull/52308)
Previously we only searched the configured instance. We now additionally
search sourcegraph.com which contains the public corpus of code we
actually want to be matching against.

This introduced a more noticeable lag which I need to investigate. These
queries should be fast on dotcom.

This isn't the final design for how this works. We will explore doing
federation on the instance to avoid anonymous requests. Additionally we
don't want to use the private instance for filtering. However, this was
quick to implement to unblock testing the corpus.

Test Plan: configured cody against s2 and did an attribution search
against the zoekt repo. It only found sourcegraph/zoekt results. Then
used this PR and it also found google/zoekt results.