Cody completions: Add unit tests (https://github.com/sourcegraph/sourcegraph/pull/52513)
This was deadly necessary: An easy way for us to write tests for the
different heuristics we're playing around with.

There's quite a bit of initial mocking involved I’m afraid but the
individual test cases are kept very clean.

My idea here is that the mocks might need to evolve but the number of
abstraction should not. It's always the same pattern:

- We have some code as an input. We use `|` to denote the current
position. (we later might have different languages as well when we start
to extend multi-line completion support or use treesitter)
- We _can_ provide a list of mocked backend responses. These can be used
to capture all the oddities we face, e.g. the "odd indentation" bug.
They are optional because for some tests we do not care about the result
(in which case we return an empty backend response
- We can then assert on the VS Code inline suggestion returned or the
requests being bade. The requests are important to have full control
over the tests (e.g. do we request a single-line or a multi-line
completion?)
- For now, this does not include any context yet but this might be added
with mocks.

Some tests are skipped right now because the fix for that is in https://github.com/sourcegraph/sourcegraph/pull/52497
and not merged yet.