Cody completions: Fix interop between spaces and tabs (https://github.com/sourcegraph/sourcegraph/pull/52497)
This PR fixes two issues related to spaces/tabs interop:

1. **Odd indentation** (= when Cody inserts a sole space at the
beginning of an indented code line) is inserted after a current eventual
tab indentation (so it looks like this: `\t\t something`). To account
for that, the regex now captures preceeding tabs and keeps them in tact.
2. **Tab to space converstion**. Today Cody started to respond with tab
indented text (for no reason?). I want to leave it to the IDE to conver
this but we have to account for this in the indentation function (where
we cound the indentaiton for a line).
 
## Test plan

- Manually verified
- I really need to add the test suite for this. Will come very very soon
I promise 🤞

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

---------

Co-authored-by: Valery Bugakov <skymk1@gmail.com>