JetBrains plugin: merge Sourcegraph and Cody (#52706)
Previously, we had two separate JetBrains plugins: Sourcegraph and Cody.
This PR merges the two plugins so that there will only be one plugin:
Sourcegraph.

This PR doesn't delete the Cody plugin yet. We can keep the
cody-jetbrains directory around for a while longer just in case we want
to consult the code side-by-side.

This PR replaces the `cody.config` and `cody.telemetry` packages in
favor of the equivalent packages in the Sourcegraph plugin. This results
in minor UX gaps that we can recover in the future with separate PRs.

## Test plan

- Green CI
- Follow CONTRIBUTING.md and verify that chat + completions work as
expected

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->
Former-commit-id: ccfdbc4a700813ff67a906f53d31c9a21a286983