Cody: Recipes PR description (https://github.com/sourcegraph/sourcegraph/pull/51721)
This PR is for issue https://github.com/sourcegraph/sourcegraph/pull/50526. The idea is to check the templates first
with most of the possible locations of the PR template stored. If it
exists, use it by summarizing the recent changes of the current checkout
branch. Else, just fallback to the summarisation of commits and a
checklist at the bottom.

Some of the sample PR templates generated:

![iScreen Shoter - Code -
230510201407](https://github.com/sourcegraph/sourcegraph/assets/44617923/c4a50630-ad92-48e4-a50a-48b2c3f0af1a)

![iScreen Shoter - Code -
230510201722](https://github.com/sourcegraph/sourcegraph/assets/44617923/e7a92a79-5571-4cac-ad46-fee300485ffa)

## Test plan
Tested it manually for different repositories with a PR template.

---------

Co-authored-by: Thorsten Ball <mrnugget@gmail.com>