cody: add `Debugging with Chrome Devtools` docs (https://github.com/sourcegraph/sourcegraph/pull/52905)
Being new to VSCode extension development, I found it wasn't obvious how
to utilize Chrome DevTools for background process debugging. It might be
known info, but it's helpful to have it documented, especially for those
new to this area or needing a refresher.

What's cool about the setup:

1. **Start from CLI**: using the newly added npm script, we can start
the extension host from the CLI.
2. **Iteration speed**: by running the build watcher in a separate
terminal window, we can reload the already started extension host to
test recent changes without rebuilding the extension + web views from
scratch before starting the extension host (that's what the launch
command does now).
3. **Immediate File Access**: The npm script automatically opens a
workspace with a source file at a specific line and column in the
extension host to test Cody's features immediately.
4. **Advanced debugging**: though the VSCode launch script does enable
debugging mode where breakpoints and logs are available in the VSCode
UI, it doesn't come close to the flexibility of the standalone version
of Chrome DevTools.