Cody completions: Various improvements to reliability and logging (https://github.com/sourcegraph/sourcegraph/pull/52935)
Closes https://github.com/sourcegraph/sourcegraph/pull/52686

There are a bunch of small changes in this PR. Here's a summary:

- I changed the default prompt to (hopefully) reduce the number of
non-code responses. This works but mostly by gut feel. We'll need
@valerybugakov's work to make sure we make some actual progress there.
This was mostly used to reduce the number of times the completion starts
with a code block end (```) tag so it increases the requests that make
it through to the user.
- Filter out some known "bad starters" that we have collected over the
last week. E.g. where cody starts the response with `➕
console.log('hi')` for example. 🙃 This means that more requests should
make it to the user.
- Instead of throwing away responses that start with whitespace, we now
trim it instead. This should mean that more requests make it through the
user.
- Empty responses (those that we filter out in the post process logic)
are no longer forwarded to VS Code. This mainly fixes a small annoyance
where the completions widget would e.g. show `1/2` completions but the
second one being empty.
- Tweaks based on [this
research](https://sourcegraph.slack.com/archives/C05A7PPR07Q/p1685715846764189)
  - We now make 3 requests instead of 2 on empty lines
- Reduced the debouncing timeout quite significantly so that our latency
becomes a lot besser.
  - ➡ We're working simultaneously to increase the rate-limit for dotcom
- Logging:
- We now log whenever we have no proper completion to show (so we can
assess how rare this is since it happened often in internal tests)
- We now count completions as `read` when they are accepted (so that we
do never count a completion as accepted but not as suggested)
- Removed some debug logging leftover 

## Test plan

Tests still green.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->