Cody: PR description recipe fix (https://github.com/sourcegraph/sourcegraph/pull/53749)
This PR is a minor fix for the PR description recipe. As I realised
later, when you checkout from any other branch to a new branch, the `git
log` command will carry all the commits not done by the current user and
a long list of other user commits too. So, I updated the command to
consider that. I have tested these changes, and it works well now!

## Test plan

![iScreen Shoter - Code -
230620173451](https://github.com/sourcegraph/sourcegraph/assets/44617923/0a3034b3-be1c-40b9-93bc-92c8c818a1d1)