Revert "Cody: Add warning message for recipes (https://github.com/sourcegraph/sourcegraph/pull/54025)" (https://github.com/sourcegraph/sourcegraph/pull/54319)
This reverts commit d62fe3e15f620c991a05a6bab2ad83986cf95268 (cc
@deepak2431)

We noticed that there is some unexpected behavior in some recipes where
this warning is shown even if only 5 lines of code are selected. It's
probably because the context lines are truncated but we don't have time
to look into this and fix it just yet (we were supposed to release a few
hours ago 😓). We'll take a closer look after the release.

## Test plan

- It's a revert

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->