autocomplete: cut-off completions on partial suffix match (#108)
We can cut off the completion on the partial suffix match (first N characters) to prevent suggesting code duplicating the completion suffix. We want to rely on a big-enough number of symbols to avoid false positives. We do not want to match the whole string because it leads to suffix duplication.