faster download-rg.sh that runs w/o hitting the GitHub API (#68)
The download-rg.sh script downloads prebuilt ripgrep binaries.
Previously, it hit the GitHub API for a list of assets associated with a
release. This took ~1-2s. Now, it hardcodes the URL for the assets based
on the desired binaries, which should be fairly stable (and if the URL
format changes, it will error, not fail silently).