Cody: Bot response multiplexer publishes, then completes turns (#70)
The LLM interaction pushes text to the bot response multiplexer without
waiting for the publishing to complete. If the turn ends before
publishing is complete, the half-parsed buffer is echoed as-is. The net
effect is that `</selection>` closing tags from fixups are sometimes
echoed to the output.

This fixes the problem by chaining a promise for the publications
together and waits for them to complete before doing the "turn complete"
processing.

## Test plan

`cd client/cody-shared && pnpm test`

Manual test:

This is racy, but you should use Fixup from Inline Instructions, Fixup
(Experimental), etc. and check that the word `</selection>` does not
appear in the output.