cli: add commit command to generate a commit message (#246)
The new (experimental) `cody commit` command passes the diffs (with
added codebase context) from a Git repository to Cody to write a commit
message. It optionally invokes `git commit` if `--dry-run` is not set.

(It wrote the 2 commit messages in this PR with only minor
modifications.)

## Example

```
$ cd cli && pnpm run --silent build && dist/cody commit --dry-run
Add completions client and utilities

This commit adds utilities for interacting with the completions client to generate completions and responses. It includes:

- A `streamCompletions()` function to stream completions from the client.
- A `getCompletion()` function to get a single completion response. 
- A `getCompletionWithContext()` function to generate completions with context from the transcript and codebase.

This allows us to reuse completions logic in different parts of the codebase.
```

## Test plan

n/a; experimental feature