use vscode.workspace.fs.stat API for file existence checks (instead of ripgrep)
The new implementation is simpler (uses a standard VS Code extension API, no dependency on ripgrep, no exec'ing a subprocess) and faster.

The previous implementation was more lenient and applied an implicit `**/` glob prefix to each file path. For example, if a workspace contained only the file `x/a/b.txt`, then the previous implementation would report that the path `a/b.txt` exists. That leniency reduced the rate of false hallucination reports. TODO

For the query `Which files define SAML auth?` on the `sourcegraph/sourcegraph` repository (same as that used in the original benchmark at https://github.com/sourcegraph/sourcegraph/pull/52541):

- Previous vscode.workspace.findFiles implementation: 2784ms (just to confirm I repro'd the terrible slowness initially reported from this impl)
- Current ripgrep implementation introduced in https://github.com/sourcegraph/sourcegraph/pull/52541: 64ms
- New implementation (using `vscode.workspace.fs.stat`): 2ms

All operated on 5 paths in Cody's response.