allow manual triggering of completions (#285)
This line was mistakenly added, based on a misunderstanding. The
`context.triggerKind` refers to how *inline completions* were triggered,
not the *suggest widget* (the dropdown menu). It caused the `Trigger
Inline Suggestion` VS Code action to not do anything in Cody. Now, that
action works.

As for the other case (`sameLinePrefix.endsWith('.')`), we are already
suppressing Cody completions when `context.selectedCompletionInfo` is
present (which is when there is a selected item in the suggest widget),
which already covers that case more robustly.

## Test plan

Run the VS Code `Trigger Inline Suggestion` action and confirm it works.