separately cache non-empty lines with differing trailing whitespace
For example, a cached completion for `a =` should not be returned for `a = ` (with a trailing
space). We still preserve the trailing-whitespace trimming behavior for entirely empty lines to
avoid rerunning completion when the user is just indenting on a new line.

Fixes the issue reported at https://github.com/sourcegraph/cody/pull/284#pullrequestreview-1534394856.