show inline completions even when suggest widget is visible (experimental)
Previously, when the suggest widget (which shows a list of function names, variable names, etc., from language symbols) was visible, Cody did not show an inline completion.

With the new experimental `cody.autocomplete.experimental.completeSuggestWidgetSelection` VS Code user setting, now Cody will show an inline completion when the first suggest widget item is selected. Its completion is based only on the characters you have typed, not the selected completion item (but the former is typically a prefix of the former). Taking the actual suggest widget completion item's `text` into account would be a UX improvement.

In a file, type a few characters that would trigger the suggest widget to show. Confirm that Cody *also* shows an inline completion (even while the suggest widget is still visible).